requirements.txt
selenium
python-dotenv
loguru
usaddress
pytest
boto3
sqlalchemy
pytest-cov
bandit
pip-audit 